Move filter and config code
